Refactoring to static methods
