+++ backend/api/_tests/test_activity_dashboard.py
@@ -124,9 +124,9 @@ class TestMetricModel:
     def test_metrics_api_using_dynamo(self, monkeypatch, commit_activity_input, commit_activity_result, latest_commit, total_commit, total_installs, recent_installs, timeline, limit):
         monkeypatch.setattr(model, 'get_commit_activity', self._validate_args_return_value(commit_activity_input))
         monkeypatch.setattr(model, 'get_latest_commit', self._validate_args_return_value(latest_commit))
-        monkeypatch.setattr(model, 'get_total_installs', self._validate_args_return_value(total_installs))
-        monkeypatch.setattr(model, 'get_recent_installs', self._validate_args_return_value(recent_installs))
-        monkeypatch.setattr(model, 'get_timeline', self._validate_args_return_value(timeline))
+        monkeypatch.setattr(model.InstallActivity, 'get_total_installs', self._validate_args_return_value(total_installs))
+        monkeypatch.setattr(model.InstallActivity, 'get_recent_installs', self._validate_args_return_value(recent_installs))
+        monkeypatch.setattr(model.InstallActivity, 'get_timeline', self._validate_args_return_value(timeline))
 
         from api.model import get_metrics_for_plugin
         actual = get_metrics_for_plugin(PLUGIN_NAME, limit, True)
