+++ backend/api/model.py
@@ -8,7 +8,7 @@ from io import BytesIO
 from collections import defaultdict
 import pandas as pd
 
-from api.install_activity import get_total_installs, get_recent_installs, get_timeline
+from api.metrics import InstallActivity
 from utils.github import get_github_metadata, get_artifact
 from utils.pypi import query_pypi, get_plugin_pypi_metadata
 from api.s3 import get_cache, cache, write_data, get_install_timeline_data, get_latest_commit, get_commit_activity, \
@@ -550,10 +550,10 @@ def _get_usage_data(plugin: str, limit: int, use_dynamo: bool = False) -> Dict[s
     :params bool use_dyanmo: Fetch data from dynamo if True, else fetch from s3. (default= False)
     """
     if use_dynamo:
-        timeline = get_timeline(plugin, limit) if limit else []
+        timeline = InstallActivity.get_timeline(plugin, limit) if limit else []
         usage_stats = {
-            'total_installs': get_total_installs(plugin),
-            'installs_in_last_30_days': get_recent_installs(plugin)
+            'total_installs': InstallActivity.get_total_installs(plugin),
+            'installs_in_last_30_days': InstallActivity.get_recent_installs(plugin)
         }
     else:
         data = get_install_timeline_data(plugin)
